Compress entire file immediately before upload.
Right before uploading to S3, if @compression is true, use zlib's
gzip writer to compress the entire file and upload the compressed
file instead. Compression is done in 64kb increments to save memory.
We also make sure that the source file and the compressed file are
also deleted after upload.